Reorganize the code a bit.
1. Extract QueryData classes from res1d.py to query.py
2. Put the method mike1d_quantities into various.py
